+++ e2e/tests/update_cli.py
@@ -12,11 +12,7 @@ storage_path = os.path.join(test_folder_path, "storage")
 local_repo_path = os.path.join(test_folder_path, "local")
 remote_repo_path = os.path.join(test_folder_path, "remote")
 
-last_checked_storage_path = os.path.join(
-    storage_path,
-    "test",
-    "last-update-check.txt",
-)
+last_checked_storage_path = os.path.join(storage_path, "test", "last-update-check.txt")
 
 os_env_vars = {
     "HEXAGON_STORAGE_PATH": storage_path,