+++ hexagon/support/execute/action.py
@@ -1,7 +1,7 @@
 import importlib
+import os
 import subprocess
 import sys
-import os
 import time
 import traceback
 from pathlib import Path
@@ -9,14 +9,13 @@ from typing import List, Union, Dict
 
 from rich import traceback as rich_traceback
 
+from hexagon.domain import configuration
+from hexagon.domain.env import Env
 from hexagon.domain.tool import ActionTool
 from hexagon.domain.tool.execution import ToolExecutionData
-from hexagon.domain.env import Env
-from hexagon.domain import configuration
-from hexagon.support.printer import log, translator
 from hexagon.support.hooks import HexagonHooks
+from hexagon.support.printer import log
 
-_ = translator
 _command_by_file_extension = {"js": "node", "sh": "sh"}
 
 
@@ -34,8 +33,7 @@ def _execute_action(tool: ActionTool, env_args, env: Env, args, custom_tools_pat
         custom_tools_path if custom_tools_path else configuration.custom_tools_path
     )
     action_to_execute: str = tool.action
-    ext = action_to_execute.split(".")[-1]
-    script_action_command = _command_by_file_extension.get(ext)
+    script_action_command = __script_action_command(action_to_execute)
 
     if script_action_command:
         _execute_script(
@@ -54,7 +52,6 @@ def _execute_action(tool: ActionTool, env_args, env: Env, args, custom_tools_pat
         )
 
         if return_code != 0:
-            # "{executed_command} returned code: {return_code}\n"
             log.error(
                 _("error.support.execute.action.command_result_code").format(
                     executed_command=executed_command, return_code=return_code
@@ -63,33 +60,34 @@ def _execute_action(tool: ActionTool, env_args, env: Env, args, custom_tools_pat
 
             if return_code == 127:
                 log.error(
-                    "{} [bold]{}".format(
-                        _("error.support.execute.action.could_not_execute"), tool.action
+                    _("error.support.execute.action.could_not_execute").format(
+                        action=tool.action
                     )
                 )
                 log.error(_("error.support.execute.action.we_tried"))
                 log.error(
-                    "  - {} [bold]{}".format(
-                        _("error.support.execute.action.attempt_cli_custom_dir"),
-                        custom_tools_path,
+                    _("error.support.execute.action.attempt_cli_custom_dir").format(
+                        path=custom_tools_path
                     )
                 )
                 log.error(
-                    "  - {} (hexagon.actions.external)".format(
-                        _("error.support.execute.action.attempt_internal_tools")
+                    _("error.support.execute.action.attempt_internal_tools").format(
+                        package="hexagon.actions.external"
                     )
                 )
                 log.error(
-                    "  - {} (.js, .sh)".format(
-                        _("error.support.execute.action.attempt_known_script")
+                    _("error.support.execute.action.attempt_known_script").format(
+                        extensions=".js, .sh"
                     )
                 )
-                log.error(
-                    "  - {}".format(
-                        _("error.support.execute.action.attempt_inline_command")
-                    )
-                )
+                log.error(_("error.support.execute.action.attempt_inline_command"))
             sys.exit(1)
+
+
+def __script_action_command(action_to_execute):
+    ext = action_to_execute.split(".")[-1]
+    script_action_command = _command_by_file_extension.get(ext)
+    return script_action_command
 
 
 def _execute_python_module(
@@ -155,6 +153,9 @@ def __sanitize_args_for_command(*args: Union[List[any], Dict, Env]):
 
 
 def _load_action_module(action_id: str, custom_tools_path):
+    if not all(s and s.isidentifier() for s in action_id.split(".")):
+        return None
+
     try:
         return __load_module(action_id)
     except ModuleNotFoundError as e: