+++ hexagon/support/help.py
@@ -1,12 +1,10 @@
 from itertools import groupby
 from typing import List
 
+from hexagon.domain.cli import Cli
 from hexagon.domain.env import Env
 from hexagon.domain.tool import Tool
-from hexagon.domain.cli import Cli
-from hexagon.support.printer import log, translator
-
-_ = translator
+from hexagon.support.printer import log
 
 
 def print_help(cli_config: Cli, tools: List[Tool], envs: List[Env]):
@@ -26,26 +24,37 @@ def print_help(cli_config: Cli, tools: List[Tool], envs: List[Env]):
 
     log.info(f"[bold]{cli_config.name}", gap_end=1)
 
-    log.info("[bold][u]{}:".format(_("msg.global.envs")))
-    for env in envs:
+    log.info(_("msg.support.help.envs"))
+    for i, env in enumerate(envs):
         log.info(
             f'  {env.name + (" (" + env.alias + ")" if env.alias else ""):<60}[dim]{env.long_name or ""}'
         )
         if env.description:
-            # TODO: if env is the last one it should not print with gap
-            # the same for tools
-            log.info(f'  {"": <60}[dim]{env.description}', gap_end=1)
+            log.info(
+                f'  {"": <60}[dim]{env.description}',
+                gap_end=gap_if_last(envs, i),
+            )
+        else:
+            log.gap()
 
-    log.info("[bold][u]{}:".format(_("msg.global.tools")), gap_start=2)
+    log.info(_("msg.support.help.tools"), gap_start=2)
 
     data = sorted(tools, key=lambda t: t.type, reverse=True)
 
     for gk, g in groupby(data, lambda t: t.type):
         log.info(f"[bold]{gk}:", gap_start=1)
+        type_tools = list(g)
 
-        for tool in g:
+        for i, tool in enumerate(type_tools):
             log.info(
                 f'  {tool.name + (" (" + tool.alias + ")" if tool.alias else ""):<60}[dim]{tool.long_name or ""}'
             )
             if tool.description:
-                log.info(f'  {"": <60}[dim]{tool.description}', gap_end=1)
+                log.info(
+                    f'  {"": <60}[dim]{tool.description}',
+                    gap_end=gap_if_last(type_tools, i),
+                )
+
+
+def gap_if_last(envs, i):
+    return 1 if i + 1 < len(envs) else 0